requirements.txt
django
djangorestframework
django-bootstrap-form
markdown
django-filter
Pillow
coverage
